+++ Changelog.md
@@ -4,6 +4,7 @@
 
 * Permalink and dedicated page for every message.
 * Logo in the header is also a link to the homepage.
+* Added syntax highlighting client-side via [SyntaxHighlighter](http://alexgorbatchev.com/SyntaxHighlighter/).
 
 ## 0.1.0